% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalyser.R
\name{runAnalyser}
\alias{runAnalyser}
\title{SeqExpressionAnalyser: Sequence Expression Analyser}
\usage{
runAnalyser()
}
\value{
A Shiny App is launched for interactive
and comprehensive differential gene expression analysis of RNA-Seq data
}
\description{
SeqExpressionAnalyser enables comprehensive differential expression analysis of RNA-Seq data. It is interactive, easy to use, and reproducible.
This function launches the main application included in the package.
}
\examples{
\dontrun{
  runAnalyser()
}
}
